refactor(framework): the telefunc types depend on the leaf, not the http server
